backend/requirements.txt
# API Framework
fastapi
uvicorn[standard]
python-multipart
mangum

# Authentification
python-jose[cryptography]
passlib[argon2]

# Base de données
sqlalchemy
alembic
psycopg[binary]
supabase<2.10.0

# IA
groq

# Utilitaires
python-dotenv
httpx
websockets
pydantic
pydantic-settings
email-validator
aiosqlite
httpx
